Migrate base component and provider structures to mgt-element package (#594)
* Migrated Providers and IProvider to mgt-element

* Migrated baseComponent to mgt-element

* Migrated template helper to mgt-element

* Migrated templatedComponent to mgt-element

* Updated tsconfigs to fix debugging

* Migrated Providers and IProvider to mgt-element

* Migrated baseComponent to mgt-element

* Migrated template helper to mgt-element

* Migrated templatedComponent to mgt-element

* Updated tsconfigs to fix debugging

* Merged react work with base-component work

* Updated readme with samples guidance

* Added preview doc for mgt-element

* Fixed samples

* Fixed storybook

* Fixed typo and removed console log

* Update generated react.ts
