+++ packages/mgt-element/src/components/baseComponent.ts
@@ -5,8 +5,8 @@
  * -------------------------------------------------------------------------------------------
  */
 
-import { LitElement, property, PropertyValues } from 'lit-element';
-import { Providers } from '../Providers';
+import { LitElement, PropertyValues } from 'lit-element';
+import { Providers } from '../providers/Providers';
 
 /**
  * Defines media query based on component width
@@ -64,7 +64,7 @@ export abstract class MgtBaseComponent extends LitElement {
    * Gets the ComponentMediaQuery of the component
    *
    * @readonly
-   * @type {ComponentMediaQuery}
+   * @type {MgtElement.ComponentMediaQuery}
    * @memberof MgtBaseComponent
    */
   public get mediaQuery(): ComponentMediaQuery {
