+++ packages/mgt-element/src/index.ts
@@ -7,3 +7,11 @@
 
 export * from './IBatch';
 export * from './IGraph';
+export * from './components/baseComponent';
+export * from './components/templatedComponent';
+export * from './providers/IProvider';
+export * from './providers/Providers';
+export * from './utils/EventDispatcher';
+export * from './utils/TemplateContext';
+export * from './utils/TemplateHelper';
+export * from './utils/equals';
