+++ packages/mgt-element/src/providers/IProvider.ts
@@ -7,8 +7,9 @@
 
 import { AuthenticationProvider } from '@microsoft/microsoft-graph-client/lib/es/IAuthenticationProvider';
 import { AuthenticationProviderOptions } from '@microsoft/microsoft-graph-client/lib/es/IAuthenticationProviderOptions';
-import { Graph } from '../Graph';
+import { IGraph } from '../IGraph';
 import { EventDispatcher, EventHandler } from '../utils/EventDispatcher';
+
 /**
  * Provider Type to be extended for implmenting new providers
  *
@@ -24,7 +25,7 @@ export abstract class IProvider implements AuthenticationProvider {
    * @type {Graph}
    * @memberof IProvider
    */
-  public graph: Graph;
+  public graph: IGraph;
   private _state: ProviderState;
   private _loginChangedDispatcher = new EventDispatcher<LoginChangedEvent>();
   /**
