+++ packages/mgt-element/src/providers/Providers.ts
@@ -5,8 +5,9 @@
  * -------------------------------------------------------------------------------------------
  */
 
-import { IProvider } from './providers/IProvider';
-import { EventDispatcher, EventHandler } from './utils/EventDispatcher';
+import { IProvider } from './IProvider';
+import { EventDispatcher, EventHandler } from '../utils/EventDispatcher';
+
 /**
  * Provides implementation for acquiring the necessary access token for calling the Microsoft Graph APIs.
  *
