+++ packages/mgt-element/src/utils/TemplateContext.ts
@@ -5,10 +5,9 @@
  * -------------------------------------------------------------------------------------------
  */
 
-export * from './components/components';
-export * from './components/providers/providers';
-export * from './providers/providers';
-export * from './graph/types';
-export { TeamsHelper } from './utils/TeamsHelper';
-export { prepScopes } from './utils/GraphHelpers';
-export * from './utils/Cache';
+/**
+ * Template context object
+ */
+export interface TemplateContext {
+  [prop: string]: any;
+}
