+++ packages/mgt-element/src/utils/TemplateHelper.ts
