+++ packages/mgt-element/tsconfig.json
@@ -1,10 +1,10 @@
 {
-  "extends": "../../tsconfig.json",
+  "extends": "../../tsconfig.base.json",
   "compilerOptions": {
     "outDir": "dist",
+    "sourceRoot": "src",
     "rootDir": "src",
-    "baseUrl": "./"
+    "composite": true
   },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules"]
+  "include": ["src"]
 }
