+++ packages/mgt-react/tsconfig.json
@@ -1,14 +1,11 @@
 {
-  "extends": "../../tsconfig.json",
+  "extends": "../../tsconfig.base.json",
   "compilerOptions": {
     "outDir": "dist/es6",
     "rootDir": "src",
-    "baseUrl": "./",
-    "paths": {
-      "@microsoft/mgt": ["../mgt/src"]
-    }
+    "baseUrl": "./"
   },
   "include": ["src/**/*"],
   "exclude": ["node_modules"],
-  "references": [{ "path": "../mgt" }]
+  "references": [{ "path": "../mgt-element" }, { "path": "../mgt" }]
 }
