+++ packages/mgt/src/components/components.ts
@@ -26,5 +26,3 @@ export * from './mgt-person/mgt-person';
 export * from './mgt-tasks/mgt-tasks';
 export * from './mgt-teams-channel-picker/mgt-teams-channel-picker';
 export * from './mgt-todo/mgt-todo';
-
-export { ComponentMediaQuery } from './baseComponent';
