+++ packages/mgt/src/components/providers/mgt-msal-provider.ts
@@ -5,9 +5,8 @@
  * -------------------------------------------------------------------------------------------
  */
 
-import { customElement, LitElement, property } from 'lit-element';
-import { Providers } from '../../Providers';
-import { LoginType } from '../../providers/IProvider';
+import { customElement, property } from 'lit-element';
+import { Providers, LoginType } from '@microsoft/mgt-element';
 import { MsalConfig, MsalProvider } from '../../providers/MsalProvider';
 import { MgtBaseProvider } from './baseProvider';
 /**
