+++ packages/mgt/src/mock/MockGraph.ts
@@ -15,7 +15,7 @@ import {
   RetryHandlerOptions,
   TelemetryHandler
 } from '@microsoft/microsoft-graph-client';
-import { MgtBaseComponent } from '../components/baseComponent';
+import { MgtBaseComponent } from '@microsoft/mgt-element';
 import { Graph } from '../Graph';
 import { chainMiddleware } from '../utils/GraphHelpers';
 import { MockProvider } from './MockProvider';
