+++ packages/mgt/src/providers/MsalProvider.ts
@@ -7,8 +7,7 @@
 
 import { AuthenticationProviderOptions } from '@microsoft/microsoft-graph-client/lib/es/IAuthenticationProviderOptions';
 import { createFromProvider } from '../Graph';
-import { IProvider, LoginType, ProviderState } from './IProvider';
-
+import { IProvider, LoginType, ProviderState } from '@microsoft/mgt-element';
 import { AuthenticationParameters, AuthError, AuthResponse, Configuration, UserAgentApplication } from 'msal';
 
 /**
