+++ packages/mgt/tsconfig.json
@@ -1,15 +1,12 @@
 {
-  "extends": "../../tsconfig.json",
+  "extends": "../../tsconfig.base.json",
   "compilerOptions": {
     "outDir": "dist/es6",
+    "sourceRoot": "src",
     "rootDir": "src",
-    "baseUrl": "./",
-    "resolveJsonModule": true,
-    "paths": {
-      "@microsoft/mgt-element": ["../mgt-element/src"]
-    }
+    "composite": true
   },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "src/bundle"],
-  "references": [{ "path": "../mgt-element" }]
+  "references": [{ "path": "../mgt-element" }],
+  "include": ["src"],
+  "exclude": ["src/bundle"]
 }
