+++ samples/examples/people-picker.html
@@ -23,7 +23,7 @@
     </div>
   </body>
   <script type="module">
-    import { Providers, ProviderState } from '../../packages/mgt/dist/es6/index.js';
+    import { Providers, ProviderState } from '../../packages/mgt-element/dist/index.js';
 
     let picker1 = document.getElementById('picker1');
     let chosenArea = document.getElementById('chosenPeople');
