+++ samples/examples/person-card-extensions.html
@@ -15,7 +15,7 @@
     </mgt-person>
 
     <script type="module">
-      import { Providers } from '../../packages/mgt/dist/es6/index.js';
+      import { Providers } from '../../packages/mgt-element/dist/index.js';
 
       let mgtPerson = document.querySelector('mgt-person');
       mgtPerson.addEventListener('templateRendered', async e => {
