+++ samples/react-app/src/index.js
@@ -13,7 +13,8 @@ import '../node_modules/@webcomponents/webcomponentsjs/webcomponents-bundle.js';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
-import { Providers, MsalProvider } from '@microsoft/mgt';
+import { Providers } from '@microsoft/mgt-element';
+import { MsalProvider } from '@microsoft/mgt';
 import { MockProvider } from '@microsoft/mgt/dist/es6/mock/MockProvider';
 
 Providers.globalProvider = new MockProvider(true);
